Added class Menu and moved there all menu list.
