--- src/main/java/Basket/AddNewProduct.java
@@ -1,7 +1,0 @@
-package Basket;
-
-public class AddNewProduct {
-    public void addNewProduct(){
-
-    }
-}
