+++ src/main/java/basket/AddNewProduct.java
@@ -1,0 +1,7 @@
+package basket;
+
+public class AddNewProduct {
+    public void addNewProductToBasket(){
+
+    }
+}
