+++ src/main/java/basket/Basket.java
@@ -1,0 +1,10 @@
+package basket;
+
+import java.util.HashMap;
+
+public class Basket {
+    public Basket(){
+        HashMap<String, Integer> basket = new HashMap<>();
+    }
+
+}
