+++ src/main/java/basket/RemoveProduct.java
@@ -1,4 +1,4 @@
-package Basket;
+package basket;
 
 public class RemoveProduct {
     public void removeProductFromBasket(){
