+++ src/main/java/menu/Menu.java
@@ -1,0 +1,12 @@
+package menu;
+
+public class Menu {
+    public void menu(){
+        System.out.println("1. Check all product's");
+        System.out.println("2. Add new product to basket");
+        System.out.println("3. Remove product from basket");
+        System.out.println("4. Check what is inside basket");
+        System.out.println("5. Pay for all product inside basket");
+        System.out.println("6. End shopping");
+    }
+}
